chore: update schemas
