--- domains/dvla/src/index.ts
@@ -1,1 +1,0 @@
-export { viewDriverResponseSchema } from "./schemas/driversLicence";
