+++ platform/domains/dvla/src/schemas/domain/authenticate.ts
@@ -1,8 +1,10 @@
 import { NonEmptyString } from "@flex/utils";
 import z from "zod";
 
-export const authenticateResponseSchema = z.object({
-  "id-token": NonEmptyString,
-  apiKeyExpiry: NonEmptyString,
-  passwordExpiry: NonEmptyString,
-});
+export const authenticateResponseSchema = z
+  .object({
+    "id-token": NonEmptyString,
+    apiKeyExpiry: NonEmptyString,
+    passwordExpiry: NonEmptyString,
+  })
+  .meta({ id: "AuthenticateResponse" });
