+++ platform/domains/dvla/src/schemas/remote/drivingLicences.ts
@@ -1,7 +1,7 @@
 import z from "zod";
 
 import { commonRequestSchema } from "../common";
-import { getLicenceResponseSchema } from "../domain/drivingLicences";
+import { viewDriverResponseSchema } from "../domain/drivingLicences";
 
 export type GetLicenceRequestSchema = z.infer<typeof commonRequestSchema>;
-export type GetLicenceResponseSchema = z.infer<typeof getLicenceResponseSchema>;
+export type GetLicenceResponseSchema = z.infer<typeof viewDriverResponseSchema>;
